+++ apps/api/src/main/java/dk/treecreate/api/authentication/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package dk.treecreate.api.payload.request;
+package dk.treecreate.api.authentication.dto.request;
 
 import javax.validation.constraints.NotBlank;
