+++ apps/api/src/main/java/dk/treecreate/api/authentication/dto/response/JwtResponse.java
@@ -1,4 +1,4 @@
-package dk.treecreate.api.payload.response;
+package dk.treecreate.api.authentication.dto.response;
 
 import java.util.List;
