+++ apps/api/src/main/java/dk/treecreate/api/authentication/models/User.java
@@ -1,4 +1,4 @@
-package dk.treecreate.api.models;
+package dk.treecreate.api.authentication.models;
 
 import javax.persistence.*;
 import javax.validation.constraints.Email;
